update util-linux: add new files
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@51753 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/util-linux/sys-utils/ipcutils.h
@@ -35,11 +35,11 @@
 # define SHM_INFO	14
 struct shm_info {
 	int used_ids;
-	ulong shm_tot;		/* total allocated shm */
-	ulong shm_rss;		/* total resident shm */
-	ulong shm_swp;		/* total swapped shm */
-	ulong swap_attempts;
-	ulong swap_successes;
+	unsigned long shm_tot;		/* total allocated shm */
+	unsigned long shm_rss;		/* total resident shm */
+	unsigned long shm_swp;		/* total swapped shm */
+	unsigned long swap_attempts;
+	unsigned long swap_successes;
 };
 #endif
 
@@ -100,7 +100,7 @@ struct ipc_limits {
 	unsigned int	semvmx;		/* semaphore max value (constant) */
 
 	int		msgmni;		/* max queues system wide */
-	size_t		msgmax;		/* max size of message */
+	uint64_t	msgmax;		/* max size of message */
 	int		msgmnb;		/* default max size of queue */
 };
